requirements.txt
# Core dependencies
llama-index>=0.12.0
llama-index-vector-stores-chroma>=0.4.0
llama-index-retrievers-bm25>=0.4.0
chromadb>=0.6.0

# MCP Protocol
mcp>=1.0.0

# LLM and Embeddings
qianfan>=0.4.0
openai>=1.0.0

# Evaluation
ragas>=0.2.0
datasets>=2.0.0

# Configuration and utilities
pydantic>=2.0.0
pyyaml>=6.0
python-dotenv>=1.0.0

# Office document support (optional)
python-docx>=0.8.11
openpyxl>=3.0.0
xlrd>=2.0.0
pypdf>=3.0.0

# API Server
fastapi>=0.100.0
uvicorn>=0.20.0
nest-asyncio>=1.5.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
httpx>=0.27.0
